requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
py4j==0.10.9.2
pyspark==3.5.2
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
